NLP/requirements.txt
numpy>=1.26.0
pandas>=2.0.0
scikit-learn>=1.4.0
torch>=2.2.0
transformers>=4.40.0
accelerate>=0.29.0
sounddevice>=0.4.6
soundfile>=0.12.1
openai-whisper>=20240930
